pyproject.toml
[project]
name = "reprovide-kubo-test"
version = "0.1.0"
description = ""
authors = [
    {name = "guillaumemichel",email = "guillaume@michel.id"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "py-cid (>=0.4.0,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
